In this case I AM a WHITE-HAT hacker
Plus it will be a cool conversation starter in an interview that my website is being hosted on my Playstation3 super slim as a dedicated web-server.

It is cheaper and economically practical in my opinion.

I also have full control and is in a point of view more secure from attacks as well due to it being a local on-premises I.a.a.S I  have full access too, and is small enough to transport unlike a dedicated server rack you will find.

It is a fun challenge for me.

"We did not decide to go to the moon because it is easy but because it is hard."

You never know what hidden discoveries I will find.

A solution to creating the very CFW for the Playstation3 Super Slim, that would be something to behold.... ;)

To the  modding and homebrew community who brought PS3hen into the world, I solute you.